pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "10x-warmup"
version = "0.1.0"
description = "Warmup repository for 10xDevs.pl program - Python/Streamlit version"
authors = [
    {name = "Przeprogramowani", email = "contact@10xdevs.pl"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "ISC"}
dependencies = [
    "streamlit>=1.31.0",
    "psycopg2-binary>=2.9.9",
    "yoyo-migrations>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests", "banking"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing"
